added dotnet new3 complete subcommand
